Create mattr_accessor
In Ruby, attr_accessor is a convenient way to define both a getter and a setter method for an instance variable in a class. It automatically creates getter and setter methods for the specified instance variable, saving you from writing repetitive code. It is often used to provide external access to instance variables while encapsulating the internal implementation details of a class