Create mattr_accessor_with_modules
In Ruby, mattr_accessor is a convenience method provided by ActiveSupport, which is a part of the Ruby on Rails framework. It stands for "module attribute accessor" and is used to create class-level attributes for modules. This means that you can define attributes that are accessible at the module level, rather than the instance level like attr_accessor provides for classes.