Create cattr_accessor
n Ruby, cattr_accessor is another convenience method provided by ActiveSupport, which is a part of the Ruby on Rails framework. It stands for "class attribute accessor" and is used to create class-level attributes for Ruby classes. This means that you can define attributes that are accessible at the class level, rather than the instance level like attr_accessor provides for instances.

In this example, we've defined a class named MyClass and used cattr_accessor to create a class-level attribute called class_attribute. This attribute can be accessed and modified at the class level, and it is shared among all instances of the class.

As you can see, we can directly access and modify the class_attribute of MyClass, just like we would with a regular class-level attribute.

cattr_accessor is particularly useful when you want to store some shared data among all instances of a class, or when you need to maintain some class-level configuration settings or global variables.

Keep in mind that, similar to mattr_accessor, cattr_accessor is part of ActiveSupport and is Rails-specific. So, you'll need to include the appropriate part of ActiveSupport (e.g., require 'active_support/all') to use it outside of a Rails application. Additionally, it's important to note that cattr_accessor defines class-level attributes, which are shared across all instances of the class. If you want instance-specific attributes, you should use regular attr_accessor for that purpose.